#115 [미션] 진행중인 미션 더보기 서버 통신 구현

+++ app/src/main/java/com/dnd_8th_4_android/wery/data/remote/model/mission/ResponseMyMissionCard.kt
@@ -9,7 +9,7 @@ data class ResponseMyMissionList(
 ) : BaseResponse()
 
 data class ResponseMyMissionCard(
-    val missionId: Long? = 0,
+    val missionId: Int? = 0,
     @SerializedName("missionName") val missionTitle: String,
     val groupName: String,
     val groupImageUrl: String,
